requirements.txt
discord.py
aiohttp
flask
python-dotenv
sqlite3
requests